+++ packages/seeds-color/gulpfile.js
@@ -12,6 +12,7 @@ const ase = require('ase-utils');
 const makeDir = require('make-dir');
 
 const versions = require('@sproutsocial/seeds-utils/versions');
+const cssPropertyName = require('@sproutsocial/seeds-utils/css-property-name');
 const sassVar = require('@sproutsocial/seeds-utils/sassvar').sassVar;
 const getGulpTask = require('@sproutsocial/seeds-utils/getgulptask');
 const constantCase = require('@sproutsocial/seeds-utils/constantcase').constantCase;
@@ -32,6 +33,8 @@ gulp.task('clean', () => {
 });
 
 gulp.task('color-scss', getGulpColorTask('web', 'scss'));
+
+gulp.task('color-css', getGulpColorTask('web', 'custom-properties.css'));
 
 gulp.task('color-js', getGulpColorTask('js', 'common.js'));
 
@@ -114,6 +117,7 @@ gulp.task('color-docs', done => {
             app: upperFirst(prop.name),
             sass: sassVar(prop.package, prop.name),
             javascript: javascriptConst(prop.package, prop.name),
+            css: cssPropertyName(prop.package, prop.name),
             swift: `UIColor().${camelCase(prop.name)}()`,
             android: constantCase(prop.name),
             python: camelCase(prop.name)
@@ -131,6 +135,7 @@ gulp.task(
   gulp.series([
     'clean',
     gulp.parallel([
+      'color-css',
       'color-scss',
       'color-js',
       'color-swift',