+++ packages/seeds-border/docs/CHANGELOG.md
@@ -1,4 +1,8 @@
 ## Changelog
 
+### v0.3.0
+
+- **New** Added support for tokens as CSS custom properties. To use them, `@import seeds-border/dist/seeds-border.css` and refer to the token names in the documentation.
+
 ### v0.2.4
 - **Updated** Improved examples in documentation.