+++ packages/seeds-motion/docs/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Changelog
 
+### v0.3.0
+
+- **New** Added support for tokens as CSS custom properties. To use them, `@import seeds-motion/dist/seeds-motion.css` and refer to the token names in the documentation.
 
 ### v0.2.8