+++ packages/seeds-networkcolor/docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Changelog
 
+### v0.4.0
+
+- **New** Added support for tokens as CSS custom properties. To use them, `@import seeds-networkcolor/dist/seeds-networkcolor.css` and refer to the token names in the documentation.
+
 ### v0.3.1
 
 - **Updated** Documentation improvements.