+++ packages/seeds-space/docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Changelog
 
+### v0.3.0
+
+- **New** Added support for tokens as CSS custom properties. To use them, `@import seeds-space/dist/seeds-space.css` and refer to the token names in the documentation.
+
 ### v0.2.6
 
 - **Updated** Documentation improvements.